sort and optimize menus
