+++ include/utils/audio.hpp
@@ -1,0 +1,7 @@
+#pragma once
+namespace Utilities {
+    void disable_bg_music();
+    void enable_bg_music();
+    void disable_sfx();
+    void enable_sfx();
+}  // namespace Utilities
