+++ include/utils/link.hpp
@@ -1,0 +1,10 @@
+#pragma once
+#include "font.h"
+
+#define LINK_DEBUG_X_OFFSET 450.0f
+#define LINK_DEBUG_Y_OFFSET 200.0f
+
+namespace Utilities {
+    void show_link_debug_info(Font &font);
+    void change_tunic_color();
+}
