+++ include/utils/loading.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace Utilities {
+    void trigger_load();
+}
