+++ src/menus/any_saves_menu.cpp
@@ -4,7 +4,8 @@
 #include "menu.h"
 #include "controller.h"
 #include "fifo_queue.h"
-#include "utils.h"
+#include "utils/cursor.hpp"
+#include "utils/lines.hpp"
 #include "gorge.h"
 #include "rollcheck.h"
 #include "save_injector.h"
