+++ src/menus/practice_menu.cpp
@@ -1,7 +1,9 @@
 #include "font.h"
 #include "menu.h"
 #include "controller.h"
-#include "utils.h"
+#include "save_injector.h"
+#include "utils/cursor.hpp"
+#include "utils/lines.hpp"
 
 #define LINES 2
