+++ src/menus/settings_menu.cpp
@@ -3,7 +3,8 @@
 #include "libtp_c/include/tp.h"
 #include "controller.h"
 #include "menu.h"
-#include "utils.h"
+#include "utils/cursor.hpp"
+#include "utils/lines.hpp"
 #include <stdio.h>
 #include "log.h"
 #define LINES 5
