+++ src/menus/tools_menu.cpp
@@ -4,7 +4,8 @@
 #include "menu.h"
 #include "input_viewer.h"
 #include "controller.h"
-#include "utils.h"
+#include "utils/cursor.hpp"
+#include "utils/lines.hpp"
 #include "timer.h"
 #include "commands.h"
 #include "gorge.h"
