+++ src/save_injector.cpp
@@ -2,7 +2,6 @@
 #include "libtp_c/include/tp.h"
 #include "libtp_c/include/system.h"
 #include "log.h"
-#include "utils.h"
 #include "menu.h"
 #include <string.h>
 #include "fs.h"
